Moved pbc-specific fields into a separate mixin
