+++ pyscf/pbc/tdscf/krhf_slow_supercell.py
@@ -18,8 +18,8 @@ procedure. Several variants of TDHF are available:
    an arbitrary number of k-points and employs k-point conservation (diagonalizes matrix blocks separately).
 """
 
-from pyscf.pbc.tools import get_kconserv
-from pyscf.tdscf import common_slow, rhf_slow
+from pyscf.tdscf.common_slow import TDERIMatrixBlocks, PeriodicMFMixin
+from pyscf.tdscf import rhf_slow
 
 from pyscf.pbc.lib.kpts_helper import loop_kkk
 from pyscf.lib import logger
@@ -36,60 +36,7 @@ from itertools import product
 # * TDRHF provides a container
 
 
-def k_nocc(model):
-    """
-    Retrieves occupation numbers.
-    Args:
-        model (RHF): the model;
-
-    Returns:
-        Numbers of occupied orbitals in the model.
-    """
-    return tuple(int(i.sum() // 2) for i in model.mo_occ)
-
-
-def k_nmo(model):
-    """
-    Retrieves number of AOs per k-point.
-    Args:
-        model (RHF): the model;
-
-    Returns:
-        Numbers of AOs in the model.
-    """
-    return tuple(i.shape[1] for i in model.mo_coeff)
-
-
-def format_frozen(frozen, nmo, nk):
-    """
-    Formats the argument into a mask array of bools where False values correspond to frozen orbitals for each k-point.
-    Args:
-        frozen (int, Iterable): the number of frozen valence orbitals or the list of frozen orbitals for all k-points or
-        multiple lists of frozen orbitals for each k-point;
-        nmo (int): the total number of molecular orbitals;
-        nk (int): the total number of k-points;
-
-    Returns:
-        The mask array.
-    """
-    space = numpy.ones((nk, nmo), dtype=bool)
-    if frozen is None:
-        pass
-    elif isinstance(frozen, int):
-        space[:, :frozen] = False
-    elif isinstance(frozen, (tuple, list, numpy.ndarray)):
-        if len(frozen) > 0:
-            if isinstance(frozen[0], int):
-                space[:, frozen] = False
-            else:
-                for i in range(nk):
-                    space[i, frozen[i]] = False
-    else:
-        raise ValueError("Cannot recognize the 'frozen' argument: expected None, int or Iterable")
-    return space
-
-
-class PhysERI(common_slow.TDERIMatrixBlocks):
+class PhysERI(TDERIMatrixBlocks, PeriodicMFMixin):
 
     def __init__(self, model, frozen=None):
         """
@@ -101,37 +48,13 @@ class PhysERI(common_slow.TDERIMatrixBlocks):
             frozen (int, Iterable): the number of frozen valence orbitals or the list of frozen orbitals for all
             k-points or multiple lists of frozen orbitals for each k-point;
         """
-        super(PhysERI, self).__init__()
-        self.model = model
-        self.space = format_frozen(frozen, len(model.mo_energy[0]), len(model.kpts))
+        TDERIMatrixBlocks.__init__(self)
+        PeriodicMFMixin.__init__(self, model, frozen=frozen)
         # Phys representation
-        self.kconserv = get_kconserv(self.model.cell, self.model.kpts).swapaxes(1, 2)
         self.__full_eri_k__ = {}
         for k in loop_kkk(len(model.kpts)):
             k = k + (self.kconserv[k],)
             self.__full_eri_k__[k] = self.ao2mo_k(tuple(self.mo_coeff[j] for j in k), k)
-
-    @property
-    def mo_coeff(self):
-        return tuple(i[:, j] for i, j in zip(self.model.mo_coeff, self.space))
-
-    @property
-    def mo_energy(self):
-        return tuple(i[j] for i, j in zip(self.model.mo_energy, self.space))
-
-    @property
-    def mo_occ(self):
-        return tuple(i[j] for i, j in zip(self.model.mo_occ, self.space))
-
-    @property
-    def nocc(self):
-        """Numbers of occupied orbitals."""
-        return k_nocc(self)
-
-    @property
-    def nmo(self):
-        """Numbers of AOs per k-point."""
-        return k_nmo(self)
 
     def ao2mo_k(self, coeff, k):
         """
@@ -251,10 +174,8 @@ class PhysERI4(PhysERI):
             frozen (int, Iterable): the number of frozen valence orbitals or the list of frozen orbitals for all
             k-points or multiple lists of frozen orbitals for each k-point;
         """
-        common_slow.TDERIMatrixBlocks.__init__(self)
-        self.model = model
-        self.space = format_frozen(frozen, len(model.mo_energy[0]), len(model.kpts))
-        self.kconserv = get_kconserv(self.model.cell, self.model.kpts).swapaxes(1, 2)
+        TDERIMatrixBlocks.__init__(self)
+        PeriodicMFMixin.__init__(self, model, frozen=frozen)
 
     def __calc_block__(self, item, k):
         if self.kconserv[k[:3]] == k[3]:
