+++ pyscf/tdscf/common_slow.py
@@ -10,6 +10,8 @@ This is a helper module defining basic interfaces.
 
 from pyscf.lib import logger
 
+from pyscf.pbc.tools import get_kconserv
+
 import numpy
 from scipy.linalg import solve
 
@@ -323,9 +325,9 @@ class TDERIMatrixBlocks(TDMatrixBlocks):
         return "ab", a, b
 
 
-def format_frozen(frozen, nmo):
+def format_frozen_mol(frozen, nmo):
     """
-    Formats the argument into a mask array of bools where False values correspond to frozen orbitals.
+    Formats the argument into a mask array of bools where False values correspond to frozen molecular orbitals.
     Args:
         frozen (int, Iterable): the number of frozen valence orbitals or the list of frozen orbitals;
         nmo (int): the total number of molecular orbitals;
@@ -355,7 +357,7 @@ class MolecularMFMixin(object):
             frozen (int, Iterable): the number of frozen valence orbitals or the list of frozen orbitals;
         """
         self.model = model
-        self.space = format_frozen(frozen, len(model.mo_energy))
+        self.space = format_frozen_mol(frozen, len(model.mo_energy))
 
     @property
     def mo_coeff(self):
@@ -368,14 +370,111 @@ class MolecularMFMixin(object):
         return self.model.mo_energy[self.space]
 
     @property
+    def mo_occ(self):
+        """MO occupation numbers."""
+        return self.model.mo_occ[self.space]
+
+    @property
     def nocc(self):
         """The number of occupied orbitals."""
         return int(self.model.mo_occ[self.space].sum() // 2)
 
     @property
     def nmo(self):
-        """The total number of olecular orbitals."""
+        """The total number of molecular orbitals."""
         return self.space.sum()
+
+
+def format_frozen_k(frozen, nmo, nk):
+    """
+    Formats the argument into a mask array of bools where False values correspond to frozen orbitals for each k-point.
+    Args:
+        frozen (int, Iterable): the number of frozen valence orbitals or the list of frozen orbitals for all k-points or
+        multiple lists of frozen orbitals for each k-point;
+        nmo (int): the total number of molecular orbitals;
+        nk (int): the total number of k-points;
+
+    Returns:
+        The mask array.
+    """
+    space = numpy.ones((nk, nmo), dtype=bool)
+    if frozen is None:
+        pass
+    elif isinstance(frozen, int):
+        space[:, :frozen] = False
+    elif isinstance(frozen, (tuple, list, numpy.ndarray)):
+        if len(frozen) > 0:
+            if isinstance(frozen[0], int):
+                space[:, frozen] = False
+            else:
+                for i in range(nk):
+                    space[i, frozen[i]] = False
+    else:
+        raise ValueError("Cannot recognize the 'frozen' argument: expected None, int or Iterable")
+    return space
+
+
+def k_nocc(model):
+    """
+    Retrieves occupation numbers.
+    Args:
+        model (RHF): the model;
+
+    Returns:
+        Numbers of occupied orbitals in the model.
+    """
+    return tuple(int(i.sum() // 2) for i in model.mo_occ)
+
+
+def k_nmo(model):
+    """
+    Retrieves number of AOs per k-point.
+    Args:
+        model (RHF): the model;
+
+    Returns:
+        Numbers of AOs in the model.
+    """
+    return tuple(i.shape[1] for i in model.mo_coeff)
+
+
+class PeriodicMFMixin(object):
+    def __init__(self, model, frozen=None):
+        """
+        A mixin to support custom slices of mean-field fields.
+
+        Args:
+            model: the base model;
+            frozen (int, Iterable): the number of frozen valence orbitals or the list of frozen orbitals;
+        """
+        self.model = model
+        self.space = format_frozen_k(frozen, len(model.mo_energy[0]), len(model.kpts))
+        self.kconserv = get_kconserv(self.model.cell, self.model.kpts).swapaxes(1, 2)
+
+    @property
+    def mo_coeff(self):
+        """MO coefficients."""
+        return tuple(i[:, j] for i, j in zip(self.model.mo_coeff, self.space))
+
+    @property
+    def mo_energy(self):
+        """MO energies."""
+        return tuple(i[j] for i, j in zip(self.model.mo_energy, self.space))
+
+    @property
+    def mo_occ(self):
+        """MO occupation numbers."""
+        return tuple(i[j] for i, j in zip(self.model.mo_occ, self.space))
+
+    @property
+    def nocc(self):
+        """The number of occupied orbitals."""
+        return k_nocc(self)
+
+    @property
+    def nmo(self):
+        """The total number of molecular orbitals."""
+        return k_nmo(self)
 
 
 def eig(m, driver=None, nroots=None, half=True):
